Stubbed out and implemented HealthGoal in Models.
This file will be used as the model for future health goal tracking.
Users will be able to choose a goal they want to track through the UI and then an object of HealthGoal will be created to track said goal.
Users will be able to create and track multiple goals through the app.